+++ database-reset/package.json
@@ -12,6 +12,8 @@
   "license": "ISC",
   "dependencies": {
     "@prisma/client": "^4.4.0",
+    "bcrypt": "^5.1.0",
+    "chance": "^1.1.9",
     "cross-fetch": "^3.1.5",
     "dotenv": "^16.0.3"
   }
